LTBL: fix normal / specular mapping if light is in negative y coordinates (relative to window)

+++ LTBL2/src/LightPointEmission.cpp
@@ -172,7 +172,7 @@ void LightPointEmission::render(const sf::View& view,
 	if (normalsEnabled) 
 	{
 		auto oglLightPosition = lightTempTexture.mapCoordsToPixel(mSprite.getPosition(), view);
-		normalsShader.setUniform("lightPosition", sf::Glsl::Vec3(static_cast<float>(oglLightPosition.x), static_cast<float>(lightTempTexture.getSize().y - oglLightPosition.y), 0.15f));
+		normalsShader.setUniform("lightPosition", sf::Glsl::Vec3(static_cast<float>(oglLightPosition.x), static_cast<float>(static_cast<int>(lightTempTexture.getSize().y) - oglLightPosition.y), 0.15f));
 
 		const auto& lightColor = mSprite.getColor();
 		normalsShader.setUniform("lightColor", sf::Glsl::Vec3(lightColor.r / 255.f, lightColor.g / 255.f, lightColor.b / 255.f));
@@ -187,7 +187,7 @@ void LightPointEmission::render(const sf::View& view,
 		lightTempTexture.draw(mSprite, &normalsShader);
 
 
-		specularShader.setUniform("lightPos", sf::Glsl::Vec3(static_cast<float>(oglLightPosition.x), static_cast<float>(lightTempTexture.getSize().y - oglLightPosition.y), 0.15f));
+		specularShader.setUniform("lightPos", sf::Glsl::Vec3(static_cast<float>(oglLightPosition.x), static_cast<float>(static_cast<int>(lightTempTexture.getSize().y) - oglLightPosition.y), 0.15f));
 		specularShader.setUniform("lightSize", sf::Glsl::Vec2(oglLightWidth, oglLightHeight));
 		specularShader.setUniform("lightColorTint", sf::Glsl::Vec3(lightColor.r / 255.f, lightColor.g / 255.f, lightColor.b / 255.f));
